Removed matching dates
My postfix on FreeBSD has the date format "Sep  1 00:00:27" while date '+%b %e %H' results in "Sep  1 18". It includes the year. Since grepping is not a precise parsing anyway, it should not matter to remove the date and leave a space before "blocking". It is an heuristic approach anyway. If log file rotation is in place, it should not matter as the value is a gauge value. If there is no log rotation, it should not matter either.